+++ seacatauth/authz/role/service.py
@@ -48,11 +48,15 @@ class RoleService(asab.Service):
 		self, tenant: Optional[str] = None, page: int = 0, limit: int = None, *,
 		resource: str = None,
 		active_only: bool = False,
+		exclude_global: bool = False,
 	):
 		collection = self.StorageService.Database[self.RoleCollection]
 		query_filter = {}
 		if tenant is not None:
-			query_filter["tenant"] = {"$in": [tenant, None]}
+			if exclude_global:
+				query_filter["tenant"] = {"$in": [tenant]}
+			else:
+				query_filter["tenant"] = {"$in": [tenant, None]}
 		if resource is not None:
 			query_filter["resources"] = resource
 		if active_only is True: