+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 ### Fix
 - Improve last login search performance (#173, PLUM Sprint 230324)
+- M2M session now has access to all the M2M credentials' assigned tenants (#186, PLUM Sprint 230324)
+- Fix tenant check in role assignment (#187, PLUM Sprint 230324)
 
 ### Features
 - Per-client configurable authorization, login and cookies (#156, PLUM Sprint 230324)