+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Improve last login search performance (#173, PLUM Sprint 230324)
 - M2M session now has access to all the M2M credentials' assigned tenants (#186, PLUM Sprint 230324)
 - Fix tenant check in role assignment (#187, PLUM Sprint 230324)
+- Fix credential service lookup (#192, PLUM Sprint 230412)
 
 ### Features
 - Per-client configurable authorization, login and cookies (#156, PLUM Sprint 230324)