+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Dropped support for custom cookie domains in the configuration (#156, PLUM Sprint 230324)
 - External login status messages changed (#185, PLUM Sprint 230324)
 - Bulk-unassign tenants using "UNASSIGN-TENANT" (#189, PLUM Sprint 230324)
+- Resource "authz:tenant:admin" is deprecated and replaced by several resources (#183, PLUM Sprint 230412)
+- Viewing and browsing all tenants requires superuser privileges (#183, PLUM Sprint 230412)
+- Seacat Admin built-in resources are not editable (#183, PLUM Sprint 230412)
 
 ### Fix
 - Improve last login search performance (#173, PLUM Sprint 230324)
@@ -21,6 +24,7 @@
 ### Features
 - Per-client configurable authorization, login and cookies (#156, PLUM Sprint 230324)
 - External login ident stored (#185, PLUM Sprint 230324)
+- Granular access control for Admin API (#183, PLUM Sprint 230412)
 
 ### Refactoring
 - OpenAPI docs updated (#181, PLUM Sprint 230324)